refactor(tests): move default integration tests directly under integration/ and remove default subdirectory
