--- airbyte-integrations/connectors/source-stripe/unit_tests/integration/default/__init__.py
@@ -1,1 +1,0 @@
-# Default integration tests
