+++ airbyte-integrations/connectors/source-stripe/unit_tests/integration/test_application_fees.py
