+++ airbyte-integrations/connectors/source-stripe/unit_tests/integration/test_charges.py
