+++ airbyte-integrations/connectors/source-stripe/unit_tests/integration/test_external_account_bank_accounts.py
