+++ airbyte-integrations/connectors/source-stripe/unit_tests/integration/test_payout_balance_transactions.py
